+++ class/actions_subtotal.class.php
@@ -2966,9 +2966,12 @@ class ActionsSubtotal
 						 }
 						if (TSubtotal::isTitle($line)) {
 							//Folder for expand
-							print ' <a class="collapse_bom" id="collapse-' . $line->id . '" href="#">';
+							$titleAttr = ($line->array_options['options_hideblock'] == 1) ? $langs->trans("Subtotal_Show") : $langs->trans("Subtotal_Hide");
+
+
+							print ' <span title="'.dol_escape_htmltag($titleAttr).'" class="collapse_bom" data-title-line-target="' . $line->id . '" id="collapse-' . $line->id . '" >';
 							print (($line->array_options['options_hideblock'] == 1) ? img_picto('', 'folder') : img_picto('', 'folder-open'));
-							print '</a>';
+							print '</span>';
 						}
 
 
@@ -3579,8 +3582,11 @@ class ActionsSubtotal
 				$TSubNc[$l->id] = (int) $l->array_options['options_subtotal_nc'];
 			}
 
+			print '<script type="text/javascript" src="'.dol_buildpath('subtotal/js/subtotal.lib.js', 1).'"></script>';
+
 			$form = new Form($db);
 			?>
+
 			<script type="text/javascript">
 				$(function() {
 					var subtotal_TSubNc = <?php echo json_encode($TSubNc); ?>;
@@ -3717,6 +3723,10 @@ class ActionsSubtotal
 	        $ThtmlData['data-issubtotal'] = 'freetext';
 	    }
 
+		$ThtmlData['data-folder-status'] = 'open';
+		if(!empty($line->array_options['options_hideblock'])){
+			$ThtmlData['data-folder-status'] = 'closed';
+		}
 
 	    // Change or add data  from hooks
 	    $parameters = array_replace($parameters , array(  'ThtmlData' => $ThtmlData )  );
@@ -4088,211 +4098,262 @@ class ActionsSubtotal
 						if ($line->array_options['options_hideblock']) $TBlocksToHide[] = $line->id;
 					}
 				}
-					?>
-						<script type="text/javascript">
-							$(document).ready(function(){
 
-								//A chaque chargement de la page, on cache toutes les lignes qui font parties des blocs à cacher (sauf les titres et sous-totaux)
-								<?php if(!empty($TBlocksToHide)) { foreach($TBlocksToHide as $id_linetitle){ ?>
-								var element = $("#collapse-<?php echo $id_linetitle ?>");
-								folderManage(element);
-								<?php }} ?>
-
-								//Lors du clic sur un dossier, on cache ou faire apparaitre les lignes contenues dans le bloc concerné
-								$(".collapse_bom").click(function() {
-									folderManage_click($(this));
-									return false;
-								});
-
-								allFolderManage();
-
-								//Fonction qui permet la gestion de l'affichage des blocs à chaque chargement de page
-								function folderManage(element){
-									//On récupère le titre concerné
-									var id_line_title = element.attr('id').replace('collapse-', '');
-
-									//on récupère les lignes concernées en fonction du titre
-									$.ajax({
-										url: '<?php echo dol_buildpath('/subtotal/script/interface.php', 1); ?>'
-										,type: 'POST'
-										,data: {
-											json:1
-											,get: 'getLinesFromTitle'
-											,element: '<?php echo $element; ?>'
-											,elementid: '<?php echo $id; ?>'
-											,lineid: id_line_title
-										}
-									}).done(function(data) {
-
-										let TSubLines = [];
-										var data = JSON.parse(data);
-										//en fonction des lignes concernées, on définit les #id (html) concernés
-										$.each(data, function (title, tlines) {
-											$.each(tlines, function (key, line) {
-												TSubLines.push('#row-' + line);
-											});
-										});
-
-										//pour chaque #id (html) concerné, on cache la ligne
-										$.each(TSubLines, function(key, value){
-											$(value).hide();
-										});
-									});
-
-									//on met à jour l'icône "dossier" pour indiquer qu'il est fermé
-									element.html('<?php echo dol_escape_js(img_picto('', 'folder')); ?>');
-								}
+				$jsConf = array(
+					'linesToHide' => $TBlocksToHide,
+					'interfaceUrl'=> dol_buildpath('/subtotal/script/interface.php', 1),
+					'element' => $element,
+					'element_id' => $id,
+					'img_folder_closed' => img_picto('', 'folder'),
+					'img_folder_open' => img_picto('', 'folder-open'),
+					'langs' => array(
+						'Subtotal_HideAll' => $langs->trans("Subtotal_HideAll"),
+						'Subtotal_ShowAll' => $langs->trans("Subtotal_ShowAll"),
+						'Subtotal_Hide' => $langs->trans("Subtotal_Hide"),
+						'Subtotal_Show' => $langs->trans("Subtotal_Show")
+					)
+				);
 
 
-								//Fonction qui permet la gestion des blocs cachés ou non lors du clic sur l'icône "dossier"
-								function folderManage_click(element) {
-									//On récupère le titre concerné
-									var id_line_title = element.attr('id').replace('collapse-', '');
+				print '<script type="text/javascript" src="'.dol_buildpath('subtotal/js/subtotal.lib.js', 1).'"></script>';
 
-									//en fonction de l'état de l'icône "folder" on détermine si l'action demandée est de caché le dossier ou de montrer le dossier
-									if(element.html().indexOf('folder-open') <= 0){
-										var action_hide = 0;
-									} else {
-										var action_hide = 1;
+				?>
+				<style>
+					.toggle-all-folder-status, .collapse_bom{
+						cursor: pointer;
+					}
+					.toggle-all-folder-status:hover, .collapse_bom:hover{
+						color: var(--colortextlink, rgb(10, 20, 100));
+					}
+				</style>
+				<script type="text/javascript">
+					$(document).ready(function(){
+						// Utilisation d'une sorte de namespace en JS
+						subtotalFolders = {};
+						(function(o) {
+							o.config = <?php print json_encode($jsConf); ?> ;
+
+							/**
+							 * Dolibarr token
+							 * @type {string}
+							 */
+							o.newToken = '';
+
+
+							/**
+							 *
+							 * @param {int} titleId
+							 * @param toggleStatus : open, closed
+							 */
+							o.toggleChildFolderStatusDisplay = function(titleId, toggleStatus = 'open'){
+								let $titleLine = $('#row-' + titleId);
+
+								let collapseBtn = $('#collapse-' + titleId);
+								if($titleLine.length>0){
+									$titleLine.attr('data-folder-status', toggleStatus);
+
+									if(toggleStatus == 'closed') {
+										collapseBtn.html(o.config.img_folder_closed);
+										collapseBtn.attr('title', o.config.langs.Subtotal_Show);
+									}else{
+										collapseBtn.html(o.config.img_folder_open);
+										collapseBtn.attr('title', o.config.langs.Subtotal_Hide);
 									}
 
-									//on récupère les lignes concernées en fonction du titre
-									$.ajax({
-										url: '<?php echo dol_buildpath('/subtotal/script/interface.php', 1); ?>'
-										,type: 'POST'
-										,data: {
-											json:1
-											,get: 'getLinesFromTitle'
-											,element: '<?php echo $element; ?>'
-											,elementid: '<?php echo $id; ?>'
-											,lineid: id_line_title
+									let childrenList = getSubtotalTitleChilds($titleLine); // renvoi la liste des id des enfants
+									childrenList.forEach((childLineId) => {
+										if(toggleStatus == 'closed'){
+											$('#'+childLineId).hide();
+										}else{
+											$('#'+childLineId).show();
 										}
-									}).done(function(data) {
-
-										var data = JSON.parse(data);
-
-										//en fonction des lignes concernées, on définit les #id (html) concernée
-										$.each(data, function (title, tlines) {
-
-											if(action_hide == 0 && title !== id_line_title){
-												//cas où le bloc fermé et contient des sous-blocs : si on ouvre le bloc parent, on ouvre pas les sous-blocs
-											} else {
-												$.each(tlines, function (key, line) {
-													//si le dossier est fermé, et qu'on clique, alors on ouvre le bloc
-													if(action_hide == 0) {
-														//pour chaque #id (html) concerné, on affiche la ligne
-														$('#row-' + line).show();
-
-														//on met à jour l'extrafield "hideblock" pour indiquer que le bloc de cette ligne doit être affiché
-														$.ajax({
-															url: '<?php echo dol_buildpath('/subtotal/script/interface.php', 1); ?>'
-															, type: 'POST'
-															, data: {
-																json: 1
-																, set: 'update_hideblock_data'
-																, lineid: title
-																, element: '<?php echo $element; ?>'
-																, elementid: '<?php echo $id; ?>'
-																, value: 0
-															}
-														})
-
-														//On met à jour l'icône "dossier"
-														$('#collapse-' + title).html('<?php echo dol_escape_js(img_picto('', 'folder-open')); ?>');
-													}
-													//si le dossier est ouvert, et qu'on clique, alors on ferme le bloc
-													else {
-														//pour chaque #id (html) concerné, on affiche la ligne
-														$('#row-' + line).hide();
-
-														//on met à jour l'extrafield "hideblock" pour indiquer que le bloc de cette ligne doit être caché
-														$.ajax({
-															url: '<?php echo dol_buildpath('/subtotal/script/interface.php', 1); ?>'
-															, type: 'POST'
-															, data: {
-																json: 1
-																, set: 'update_hideblock_data'
-																, lineid: title
-																, element: '<?php echo $element; ?>'
-																, elementid: '<?php echo $id; ?>'
-																, value: 1
-															}
-														})
-
-														//On met à jour l'icône "dossier"
-														$('#collapse-' + title).html('<?php echo dol_escape_js(img_picto('', 'folder')); ?>');
-													}
-												});
-											}
-										});
-
 									});
 								}
+							}
 
 
-								//Fonction qui permet d'ajouter l'option "Cacher les lignes" ou "Afficher les lignes"
-								function allFolderManage(){
+							if(o.config.linesToHide && o.config.linesToHide.length > 0){
+								o.config.linesToHide.forEach((lineId) => {
+									o.toggleChildFolderStatusDisplay(lineId, 'closed');
+								});
+							}
 
-									//On ajoute une ligne au début du tableau qui propose de cacher toutes les lignes ou des les afficher
-									$('#tablelines>tbody:first').prepend('<tr><td colspan="100%" style="  text-align:right "><a id="hide_all" href="#"><?php echo img_picto('', 'folder-open', 'class="paddingright"').$langs->trans("Subtotal_HideAll"); ?></a>&nbsp;<a id="show_all" href="#"><?php echo img_picto('', 'folder', 'class="paddingright"').$langs->trans("Subtotal_ShowAll") ?></a></td></tr>')
 
-									//Lorsqu'on clique sur "cacher les lignes"
-									$("#hide_all").click(function() {
-										//on cache toutes les lignes de l'objet sauf les titres et sous-totaux
-										$("tr[data-product_type='0']").hide();
-										$("tr[data-product_type='1']").hide();
+							// Lors du clic sur un dossier, on cache ou faire apparaitre les lignes contenues dans le bloc concerné
+							$(document).on("click",".collapse_bom",function(event) {
+								event.preventDefault();
+								let targetTitleLineId = $(this).attr('data-title-line-target');
+								if(targetTitleLineId != undefined){
+									// folderManage_click(targetTitleLineId);
+									let titleRow = $('#row-' + targetTitleLineId);
+									let newStatus = titleRow.attr('data-folder-status') == 'closed' ? 'open' : 'closed'
+									let sendData = {
+										element : o.config.element,
+										element_id : o.config.element_id,
+										titleStatusList : [{
+											'id': targetTitleLineId,
+											'status': newStatus !== 'closed' ? 0 : 1,
+										}]
+									};
 
-										//on change tous les icônes "dossier" en fermé
-										$("[id*=collapse-]").html('<?php echo dol_escape_js(img_picto('', 'folder')); ?>');
-
-										//on sauvegarde l'information "hideblock" pour toutes les sections de l'objet
-										$.ajax({
-											url: '<?php echo dol_buildpath('/subtotal/script/interface.php', 1); ?>'
-											, type: 'POST'
-											, data: {
-												json: 1
-												, set: 'updateall_hideblock_data'
-												, element: '<?php echo $element; ?>'
-												, elementid: '<?php echo $id; ?>'
-												, value: 1
-											}
-										})
-										return false;
-									});
-
-									//Lorsqu'on clique sur "afficher les lignes"
-									$("#show_all").click(function() {
-										//on affiche toutes les lignes de l'objet sauf les titres et sous-totaux
-										$("tr[data-product_type='0']").show();
-										$("tr[data-product_type='1']").show();
-
-										//on change tous les icônes "dossier" en ouvert
-										$("[id*=collapse-]").html('<?php echo dol_escape_js(img_picto('', 'folder-open')); ?>');
-
-										//on sauvegarde l'information "hideblock" pour toutes les sections de l'objet
-										$.ajax({
-											url: '<?php echo dol_buildpath('/subtotal/script/interface.php', 1); ?>'
-											, type: 'POST'
-											, data: {
-												json: 1
-												, set: 'updateall_hideblock_data'
-												, element: '<?php echo $element; ?>'
-												, elementid: '<?php echo $id; ?>'
-												, value: 0
-											}
-										})
-
-										return false;
-									});
+									o.callInterface('set' , 'update_hideblock_data', sendData, function(response){
+										o.toggleChildFolderStatusDisplay(targetTitleLineId, newStatus);
+									})
 								}
-
 							});
-						</script>
+
+
+							//Fonction qui permet d'ajouter l'option "Cacher les lignes" ou "Afficher les lignes"
+							$('#tablelines>tbody:first').prepend(
+								'<tr>' +
+								'	<td colspan="100%" style="  text-align:right ">' +
+								'		<span id="hide_all"  class="toggle-all-folder-status" data-folder-status="closed" >'+o.config.img_folder_open+'&nbsp;'+o.config.langs.Subtotal_HideAll+'</span>' +
+								'		&nbsp;' +
+								'		<span id="show_all" class="toggle-all-folder-status" data-folder-status="open"  >'+o.config.img_folder_closed + '&nbsp;'+o.config.langs.Subtotal_ShowAll+'</span>' +
+								'	</td>' +
+								'</tr>'
+							);
+
+
+							// Lors du clic sur un dossier, on cache ou faire apparaitre les lignes contenues dans le bloc concerné
+							$(document).on("click",".toggle-all-folder-status",function(event) {
+								event.preventDefault();
+								newStatus = $( this ).attr('data-folder-status');
+								$( this ).fadeOut();
+
+								let sendData = {
+									element : o.config.element,
+									element_id : o.config.element_id,
+									titleStatusList : []
+								};
+
+								$('#tablelines tr[data-issubtotal=title]').each(function( index ) {
+									sendData.titleStatusList.push({
+										'id': $( this ).attr('data-id'),
+										'status': newStatus !== 'closed' ? 0 : 1,
+									});
+
+									//TODO manage response feedback to rollback display on error
+									o.toggleChildFolderStatusDisplay($( this ).attr('data-id'), newStatus);
+								});
+
+								o.callInterface('set' , 'update_hideblock_data', sendData, function(response){
+									// $('#tablelines tr[data-issubtotal=title]').each(function( index ) {
+									// 	//TODO manage response feedback
+									// });
+								});
+
+
+								$( this ).fadeIn();
+							});
+
+
+							/**
+							 *
+							 * @param {string} typeAction
+							 * @param {string} action
+							 * @param sendData
+							 * @param callBackFunction
+							 */
+							o.callInterface = function ( typeAction = 'get' , action, sendData = {}, callBackFunction){
+
+								let ajaxData = {
+									'data': sendData,
+									'token': o.newToken,
+								};
+
+								if(typeAction == 'set'){
+									ajaxData.set = action;
+								}else{
+									ajaxData.get = action;
+								}
+
+								$.ajax({
+									method: 'POST',
+									url: o.config.interfaceUrl,
+									dataType: 'json',
+									data: ajaxData,
+									success: function (response) {
+
+										if (typeof callBackFunction === 'function'){
+											callBackFunction(response);
+										} else {
+											console.error('Callback function invalide for callKanbanInterface');
+										}
+
+										if(response.newToken != undefined){
+											o.newToken = response.newToken;
+										}
+
+										if(response.msg.length > 0) {
+											o.setEventMessage(response.msg, response.result > 0 ? true : false, response.result == 0 ? true : false );
+										}
+									},
+									error: function (err) {
+
+										if(err.responseText.length > 0){
+
+											// detect login page in case of just disconnected
+											let loginPage = $(err.responseText).find('[name="actionlogin"]');
+											if(loginPage != undefined && loginPage.val() == 'login'){
+												o.setEventMessage(o.langs.errorAjaxCallDisconnected, false);
+
+												setTimeout(function (){
+													location.reload();
+												}, 2000);
+
+											}else{
+												o.setEventMessage(o.langs.errorAjaxCall, false);
+											}
+										}
+										else{
+											o.setEventMessage(o.langs.errorAjaxCall, false);
+										}
+									}
+								});
+							}
+
+
+							/**
+							 *
+							 * @param {string} msg
+							 * @param {boolean} status
+							 * @param {boolean} sticky
+							 */
+							o.setEventMessage = function (msg, status = true, sticky = false){
+
+								let jnotifyConf = {
+									delay: 1500                               // the default time to show each notification (in milliseconds)
+									, type : 'error'
+									, sticky: sticky                             // determines if the message should be considered "sticky" (user must manually close notification)
+									, closeLabel: "&times;"                     // the HTML to use for the "Close" link
+									, showClose: true                           // determines if the "Close" link should be shown if notification is also sticky
+									, fadeSpeed: 150                           // the speed to fade messages out (in milliseconds)
+									, slideSpeed: 250                           // the speed used to slide messages out (in milliseconds)
+								}
+
+
+								if(msg.length > 0){
+									if(status){
+										jnotifyConf.type = '';
+										$.jnotify(msg, jnotifyConf);
+									}
+									else{
+										$.jnotify(msg, jnotifyConf);
+									}
+								}
+								else{
+									$.jnotify('ErrorMessageEmpty', jnotifyConf);
+								}
+							}
+
+						})(subtotalFolders);
+
+					});
+				</script>
+
 				<?php
-
-
 			}
-			/**Gestion des dossiers qui permettent de réduire un bloc**/
 
 	}
 }